requirements.txt
pillow
flask
Flask-WTF
gunicorn
